use oapi-generator for GHHooks models
