--- Lambdas/GHHooks/GithubClient.swift
@@ -1,1 +1,0 @@
-import OpenAPIAsyncHTTPClient
