+++ Lambdas/GHHooks/Models/GHEvent.swift
@@ -1,5 +1,12 @@
 import Foundation
 
+typealias Repository = Components.Schemas.repository
+typealias User = Components.Schemas.simple_user
+typealias Organization = Components.Schemas.organization_simple
+typealias PullRequest = Components.Schemas.pull_request_simple
+typealias Issue = Components.Schemas.issue
+typealias Label = Components.Schemas.label
+
 struct GHEvent: Codable {
     let action: String?
     let sender: User
