--- Lambdas/GHHooks/Models/Permissions.swift
@@ -1,5 +1,0 @@
-
-struct Permissions: Codable {
-    let admin, maintain, pull, push: Bool?
-    let triage: Bool?
-}
