+++ Lambdas/GHHooks/openapi-generator-config.yml
@@ -1,3 +1,3 @@
 generate:
   - types
-  - client
+#  - client
